chore: no more v2 for user-operator related services
